refactor(cars): add cars class
Cars 클래스 추가 적용
